Avoid producing invalid iterators in division loops (#309)
This decouples loop end expression from the iterator decrement.
Previously, the `it--` produced invalid iterator to `u[-1]`
in the last iteration. It might have not been a big deal for execution
because the loop terminated and the iterator was unused,
but this was theoretical undefined behavior and constexpr context
didn't like it either.